backend/requirements.txt
flask
flask-cors
requests
faiss-cpu
numpy
google-generativeai
sentence-transformers
python-dotenv
